Updated Code

+++ fresh_analysis.sh
@@ -1,3 +1,12 @@
 #!/bin/bash
-cd ./output_files/
-rm -r *.*
+if [ -d "./output_files" ]; then
+    echo "-----------------------------------------------------"
+else
+    mkdir output_files
+    echo "Creating directory for output files"
+fi
+# DIR = /output_files
+if [ "$(ls -A ./output_files)" ]; then
+    cd ./output_files/
+    rm -r *.*
+fi
